Adds helpers.front.parseBody
