+++ lib/middleware/messages/support/params.js
@@ -1,33 +1,20 @@
 'use strict';
 
 const front = require('../../../front');
-const logger = require('../../../logger');
 const helpers = require('../../../helpers');
 
-module.exports = function supportParams() {
+module.exports = function params() {
   return (req, res, next) => {
-    const body = req.body;
-
     if (!front.isValidRequest(req)) {
       return helpers.sendResponseWithStatusCode(res, 401, 'X-Front-Signature is not valid.');
     }
-
-    helpers.request.setPlatform(req);
-    // Set template to support to indicate this was a message sent from a human support agent.
-    helpers.request.setOutboundMessageTemplate(req, 'support');
-
-    // TODO: Create helpers.front and move this to its parseBody method
-    // @see https://dev.frontapp.com/#sending-messages
-
-    helpers.request.setOutboundMessageText(req, body.text);
-    logger.debug('origin=front', { text: req.outboundMessageText }, req);
-    const recipients = body.recipients;
-    const fromRecipient = recipients.find(recipient => recipient.role === 'from');
-    req.agentId = fromRecipient.handle;
-    const toRecipient = recipients.find(recipient => recipient.role === 'to');
-    helpers.request.setUserId(req, toRecipient.handle);
-    req.platformMessageId = body.id;
-    req.frontConversationUrl = body._links.related.conversation;
+    try {
+      helpers.front.parseBody(req);
+      helpers.request.setPlatform(req);
+      helpers.request.setOutboundMessageTemplate(req, 'support');
+    } catch (err) {
+      return helpers.sendErrorResponse(res, err);
+    }
 
     return next();
   };
